+++ src/lib/modals/modification/components/ModalContentStimulusModification.svelte
@@ -4,7 +4,7 @@
   import GeneralInputText from '$lib/components/General/GeneralInput/GeneralInputText.svelte'
   import { SortableTableHeader, SectionHeader } from '$lib/modals'
   import { getStimuli, updateMultipleStimuli } from '$lib/stores/dataStore'
-  import { addErrorToast, addSuccessToast } from '$lib/stores/toastStore'
+  import { addErrorToast, addSuccessToast } from '$lib/toaster'
   import type { BaseInterpretedDataType } from '$lib/type/Data/InterpretedData/BaseInterpretedDataType'
   import { flip } from 'svelte/animate'
   import { fade } from 'svelte/transition'
