+++ src/lib/plots/scarf/types/ScarfTooltipFillingType.ts
