+++ src/lib/toaster/components/index.ts
@@ -1,0 +1,1 @@
+export { default as Toaster } from './Toaster.svelte'
