--- src/lib/type/Filling/ModalParticipantsGroup/ModalParticipantsGroupFillingType.ts
@@ -1,8 +1,0 @@
-import type { ParticipantsGroup } from '$lib/type/Data/ParticipantsGroup'
-
-export interface ModalParticipantsGroupType extends ParticipantsGroup {
-  id: number
-  name: string
-  participantsIds: number[]
-  type: 'panel' | 'editing'
-}
